home and clothing page implented
