--- apps/api/components/Footer.module.css
@@ -1,23 +1,0 @@
-.footer
-{
-    display: flex;
-    gap: 2rem;
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    padding: 1rem;
-}
-.footer a 
-{
-
-    color: black;
-    text-decoration: none;
-    font-weight: 600;
-    padding: .5rem;
-
-    text-shadow: 
-    -1px -1px 0 white,  
-     1px -1px 0 white,
-    -1px  1px 0 white,
-     1px  1px 0 white;
-}
