+++ apps/web/src/app/components/Body.module.css
@@ -1,0 +1,40 @@
+.imageContainer
+{
+    position: relative;
+    width: 100%;
+}
+
+.image
+{
+    width: 100%;
+    height: auto;
+    display: block;
+    object-fit: cover;
+}
+.footerContainer 
+{
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+.footerContainer a
+{
+    color: black;
+    text-decoration: none;
+    font-weight: 600;
+    padding: .5rem;
+
+    text-shadow: 
+    -1px -1px 0 white,  
+     1px -1px 0 white,
+    -1px  1px 0 white,
+     1px  1px 0 white;
+}
+.bodyBackground {
+  position: relative;  /* or static */
+  z-index: 0;          /* keep it behind */
+}
