+++ apps/web/src/app/components/ClothingBody.module.css
@@ -1,0 +1,142 @@
+/* Container: 4-up product grid on desktop */
+.body {
+  display: grid;
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  gap: 2rem;                 /* space between cards */
+  padding: 2rem 1rem;        /* page breathing room */
+  max-width: 1200px;         /* keeps content from stretching too wide */
+  margin: 0 auto;            /* center the grid */
+}
+
+/* Each product “card” */
+.body section {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;   /* left-align text like the screenshot */
+}
+
+/* Image fills the card width and keeps a nice aspect */
+/* ClothingBody.module.css */
+.image {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  object-position: center;
+  transition: transform 0.3s ease; /* smooth swap */
+  transform: scale(1); 
+}
+
+/* second image */
+.image2 {
+  width: 100%;
+  aspect-ratio: 4 / 5;   /* keeps every slot uniform */
+  object-fit: contain;   /* show full shirt */
+  object-position: center; /* center it */
+  background: #fff;      /* white behind transparent areas */
+  display: block;
+  border-radius: 8px;
+}
+
+/* Title & price styles */
+.description {
+  margin-top: 1rem;
+  font-size: 12px;
+  font-weight: 700;
+  color: #000;
+  letter-spacing: 0;
+
+}
+
+.price {
+  margin-top: 0;        /* remove vertical gap */
+  font-size: 12px;
+  font-weight: 600;
+  color: #000;
+  padding: 0;
+}
+
+/* Responsive: collapse to 3 / 2 / 1 columns */
+@media (max-width: 1200px) {
+  .body { grid-template-columns: repeat(3, minmax(0, 1fr)); }
+}
+@media (max-width: 900px) {
+  .body { grid-template-columns: repeat(2, minmax(0, 1fr)); }
+}
+@media (max-width: 600px) {
+  .body { grid-template-columns: 1fr; }
+}
+.card {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.imageWrapper {
+  position: relative;
+  width: 100%;
+  aspect-ratio: 4 / 5;  /* fixed slot */
+  overflow: hidden;     /* crop anything that overflows */
+  background: #fff;
+  border-radius: 8px;
+  box-shadow: 0px 0px 6px 2px rgba(0,0,0,0.15);
+}
+
+.arrow {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  background: rgba(255,255,255,0.8);
+  border: none;
+  font-size: 2rem;
+  cursor: pointer;
+
+  border-radius: 50%;
+  display: none; /* hidden by default */
+}
+
+.left { left: 8px; }
+.right { right: 8px; }
+
+.imageWrapper:hover .arrow {
+  display: block; /* show arrows on hover */
+}
+.backImage {
+  transform: scale(1);   /* zoom in slightly */
+}
+.sizeOverlay {
+  position: absolute;
+  left: 50%;
+  bottom: 10px;
+  transform: translateX(-50%);
+  padding: 6px 10px;
+  color: #000;
+  font-size: 12px;
+  font-weight: 600;
+  letter-spacing: 2px;
+  border-radius: 4px;
+  opacity: 0;                       /* hidden by default */
+  transition: opacity .25s ease;
+  z-index: 2;                       /* above the image */
+  pointer-events: none;             /* arrows still clickable */
+}
+
+.imageWrapper:hover .sizeOverlay {
+  opacity: 1;
+}
+
+.card h3, .card h4
+{
+  margin: 0;
+  letter-spacing: -1;
+}
+.card h3
+{
+  margin-top: 1.5rem;
+} 
+.card h4
+{
+  margin-bottom: 0rem;
+  font-weight: 100;
+}
