+++ apps/web/src/app/test-clothing/page.js
@@ -1,0 +1,16 @@
+import ClothingBody from '../components/ClothingBody';
+import Header from '../components/Header';
+
+function ClothingPage()
+{
+    return(
+
+        <>
+            <Header/>
+            <ClothingBody/> 
+        </>
+
+    );
+}
+
+export default ClothingPage;
