+++ apps/web/src/app/test-home/page.js
@@ -1,0 +1,15 @@
+import Header from '../components/Header';
+import Body from '../components/Body';
+
+function HomePage()
+{
+    return (
+
+        <>
+            <Header/>
+            <Body/>
+        </>
+    );
+}
+
+export default HomePage;
